Перестройка механизма подсоединения втыкал, делающая его более независимым от имени главного исполняемого модуля и его системных экспортных структур. Доделка и правка ошибок во всех втыкалах (тестовой и файловой), входящих в состав проекта. Более полная обработка некоторых маловероятных ошибок.

+++ ExternalObjectDemo/ExternalObjectDemo.cpp
@@ -56,7 +56,7 @@ namespace
         else if (holds_alternative<int>(link_val))
             return get<int>(link_val);
         else if (holds_alternative<double>(link_val))
-            return get<double>(link_val);
+            return static_cast<int>(get<double>(link_val));
         else
             return 0;
     }
